+++ modules/stream_manager/src/cpp/PrintfHelper.hpp
@@ -16,18 +16,46 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "StringPrintf.hpp"
-#include "PrintfFunction.hpp"
-#include "characters_encoding.hpp"
+#pragma once
+//=============================================================================
+#include "ArrayOf.hpp"
 //=============================================================================
 namespace Nelson {
 //=============================================================================
-bool
-StringPrintf(
-    std::wstring& result, std::wstring& errormsg, Evaluator* eval, const ArrayOfVector& arg)
+class PrintfHelper
 {
-    return printfFunction(arg, errormsg, result);
-}
+private:
+    const ArrayOfVector args;
+    int vectorIndex;
+    int elementIndex;
+    bool hasMoreData;
+    bool dataUsed;
+    void
+    IncrementDataPointer(void);
+    static int
+    flagCharacter(wchar_t c);
+    static int
+    convSpec(wchar_t c);
+    static std::wstring
+    ConvertEscapeSequences(const std::wstring& src);
+
+public:
+    PrintfHelper(const ArrayOfVector& arg_);
+    bool
+    GetNextVariableAsDouble(double& data, std::wstring& errorMessage, bool& isEmpty);
+    bool
+    GetNextVariableAsLongLong(long long& data, std::wstring& errorMessage, bool& isEmpty);
+    bool
+    GetNextVariableAsString(std::wstring& str, std::wstring& errorMessage);
+    bool
+    HasMoreData(void);
+    bool
+    WasDataUsed(void);
+    static bool
+    isEscape(wchar_t* dp);
+    static wchar_t*
+    validateFormatSpec(wchar_t* cp);
+};
 //=============================================================================
-}
+} // namespace Nelson
 //=============================================================================